25 develop a crud api for a theoretical block theory

+++ RLPortalBackend/Services/Impl/TheoryService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using RLPortalBackend.Entities;
+using RLPortalBackend.Exceptions;
 using RLPortalBackend.Models.Exercise;
 using RLPortalBackend.Models.Theory;
 using RLPortalBackend.Repositories;
@@ -18,11 +19,11 @@ namespace RLPortalBackend.Services.Impl
             _mapper = mapper;
         }
 
-        public async Task<TheoryDto> CreateAsync(CreateTheoryDto theoryDto)
+        public async Task<TheoryDto> CreateAsync(NoIdTheoryDto theoryDto)
         {
             TheoryEntity theoryEntity = _mapper.Map<TheoryEntity>(theoryDto);
 
-            IEnumerable<CreateTheorySectionDto> createTheorySectionDtos = theoryDto.CreateTheorySectionDtos;
+            IEnumerable<TheorySectionDto> createTheorySectionDtos = theoryDto.TheorySectionDtos;
             IEnumerable<TheorySectionEntity> theorySectionEntities = _mapper.Map<IEnumerable<TheorySectionEntity>>(createTheorySectionDtos);
 
             theoryEntity.TheorySectionEntities = theorySectionEntities.ToList();
@@ -33,7 +34,7 @@ namespace RLPortalBackend.Services.Impl
 
         public async Task<ICollection<TheoryDto>> GetAsync()
         {
-            IEnumerable<TheoryEntity> theoryEntities = await _repository.GetAsync();
+            ICollection<TheoryEntity> theoryEntities = await _repository.GetAsync();
             ICollection<TheoryDto> theoryDtos = _mapper.Map<ICollection<TheoryDto>>(theoryEntities);
             return theoryDtos;
 
@@ -42,18 +43,37 @@ namespace RLPortalBackend.Services.Impl
         public async Task<TheoryDto> GetByIdAsync(Guid id)
         {
             TheoryEntity theoryEntity = await _repository.GetAsync(id);
+            if (theoryEntity == null)
+            {
+                throw new TheoryNotFoundException($"Theory {id} not found");
+            }
             TheoryDto theoryDto = _mapper.Map<TheoryDto>(theoryEntity);
             return theoryDto;
         }
 
         public async Task RemoveAsync(Guid id)
         {
+            if (await _repository.GetAsync(id) == null)
+            {
+                throw new TheoryNotFoundException($"Theory {id} not found");
+            }
+
             await _repository.RemoveAsync(id);
         }
 
-        public async Task UpdateAsync(Guid id, UpdateTheoryDto updateTheoryDto)
+        public async Task UpdateAsync(Guid id, NoIdTheoryDto updateTheoryDto)
         {
-            throw new NotImplementedException();
+            if (await _repository.GetAsync(id) == null)
+            {
+                throw new TheoryNotFoundException($"Theory {id} not found");
+            }
+
+            TheoryDto dto = _mapper.Map<TheoryDto>(updateTheoryDto);
+            dto.Id = id;
+            ICollection<TheorySectionEntity> theorySectionEntities = _mapper.Map<ICollection<TheorySectionEntity>>(updateTheoryDto.TheorySectionDtos);
+            TheoryEntity newEntity = _mapper.Map<TheoryEntity>(dto);
+            newEntity.TheorySectionEntities = theorySectionEntities;
+            await _repository.UpdateAsync(id, newEntity);
         }
     }
 }
